Separated MemoryInterface
